autoload/gitcha.vim: Abbreviate commit SHA in the popup menu
Only show the first 10 characters of the commit SHA in the popup menu to
allow more room for the subject. The first part of the SHA is enough to
be able to identify it, so this gets rid of some non-essential
information.

The full commit SHA will still be inserted as a result of the
completion.