autoload/gitcha.vim: Fix match list population
Previously, we were using the completion function sample copied from the
Vim manual. This function determined `a:findstart` using a match against
`\a`, which only matches alphanumeric characters.

Git commit SHAs are SHA-1s, which can contain alphanumeric characters.
With the previous match regex, if the word to be completed contained
numbers, the `a:findstart` match would fail. This resulted in `a:base`
being empty, causing all SHAs to be added to the `matches` array.

Now we correctly match only the characters that can appear in a SHA-1,
and the completion correctly fills in the rest given the start of a SHA.